requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.2
Pillow==10.1.0
cloudinary==1.37.0
dj3-cloudinary-storage==0.0.6
urllib3==1.26.9
requests==2.26.0
django-filter==23.5
django-rest-knox==4.1.0
django-heroku==0.3.1
coreapi==2.3.3

cryptography==41.0.7

dj-rest-auth==2.2.4